// src/types/themeTypes/ThemeType.ts
export type ThemeTypes = {
  name: string;
  dark: boolean;
  variables?: object;
  colors: {
    primary?: string;
    secondary?: string;
    info?: string;
    success?: string;
    accent?: string;
    warning?: string;
    error?: string;
    lightprimary?: string;
    lightsecondary?: string;
    lightsuccess?: string;
    lighterror?: string;
    lightinfo?: string;
    lightwarning?: string;
    textPrimary?: string;
    textSecondary?: string;
    borderColor?: string;
    hoverColor?: string;
    inputBorder?: string;
    containerBg?: string;
    surface?: string;
    "on-surface-variant"?: string;
    grey100?: string;
    grey200?: string;
    muted?: string;
    "action-generic-color"?: string;
    "action-save-color"?: string;
    "action-added-color"?: string;
    "action-edit-color"?: string;
    "action-update-color"?: string;
    "action-clear-color"?: string;
    "action-return-color"?: string;
    "account-user-color"?: string;
  };
};